requirements.txt
tensorforce~=0.4.0
jupyter~=1.0
tensorflow-gpu~=1.4.0
docker~=3.1
gym~=0.10
scipy~=1.0
Pillow~=5.0
ruamel.yaml~=0.15
Flask~=0.12
requests~=2.18
